api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
requests==2.31.0